requirements.txt
fastapi
uvicorn
pydantic
supabase
langchain-groq
chromadb
sentence-transformers
huggingface-hub
httpx
langchain
langchain-community
pypdf
supabase
python-dotenv
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0